Lab 5 zrobiony zostało 6 i 7

+++ Lab05/src/FiguryGeometryczne/Trojkat.java
@@ -2,32 +2,32 @@ package FiguryGeometryczne;
 
 public class Trojkat {
 
-    int bok_a,bok_b,bok_c,h;
+    double bok_a,bok_b,bok_c,h;
 
-    public Trojkat(int bok_a, int bok_b, int bok_c, int h) {
+    public Trojkat(double bok_a, double bok_b, double bok_c, double h) {
         this.bok_a = bok_a;
         this.bok_b = bok_b;
         this.bok_c = bok_c;
         this.h = h;
     }
 
-    public int pole(){
+    public double pole(){
         return (bok_a*h)/2;
     }
 
-    public int getBok_a() {
+    public double getBok_a() {
         return bok_a;
     }
 
-    public int getBok_b() {
+    public double getBok_b() {
         return bok_b;
     }
 
-    public int getBok_c() {
+    public double getBok_c() {
         return bok_c;
     }
 
-    public int getH() {
+    public double getH() {
         return h;
     }
 }
